Refactor folders + Air Trait
