+++ src/air/mod.rs
@@ -1,3 +1,5 @@
+pub mod traits;
+pub mod cairo_air;
 pub mod constraints;
 pub mod context;
 #[cfg(debug_assertions)]
@@ -5,5 +7,3 @@ pub mod debug;
 pub mod example;
 pub mod frame;
 pub mod trace;
-pub mod cairo_air;
-pub mod air;
